+++ beacon-chain/rpc/eth/beacon/handlers_pool_test.go
@@ -2,17 +2,32 @@ package beacon
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
+	"reflect"
+	"strings"
 	"testing"
 
+	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/prysmaticlabs/go-bitfield"
+	grpcutil "github.com/prysmaticlabs/prysm/v4/api/grpc"
+	blockchainmock "github.com/prysmaticlabs/prysm/v4/beacon-chain/blockchain/testing"
+	"github.com/prysmaticlabs/prysm/v4/beacon-chain/core/signing"
+	"github.com/prysmaticlabs/prysm/v4/beacon-chain/core/transition"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/operations/attestations"
+	p2pMock "github.com/prysmaticlabs/prysm/v4/beacon-chain/p2p/testing"
+	"github.com/prysmaticlabs/prysm/v4/config/params"
+	"github.com/prysmaticlabs/prysm/v4/crypto/bls"
 	"github.com/prysmaticlabs/prysm/v4/encoding/bytesutil"
+	ethpbv1 "github.com/prysmaticlabs/prysm/v4/proto/eth/v1"
 	ethpbv1alpha1 "github.com/prysmaticlabs/prysm/v4/proto/prysm/v1alpha1"
 	"github.com/prysmaticlabs/prysm/v4/testing/assert"
 	"github.com/prysmaticlabs/prysm/v4/testing/require"
+	"github.com/prysmaticlabs/prysm/v4/testing/util"
+	"github.com/prysmaticlabs/prysm/v4/time/slots"
+	"google.golang.org/grpc"
 )
 
 func TestListAttestations(t *testing.T) {
@@ -156,4 +171,298 @@ func TestListAttestations(t *testing.T) {
 			assert.Equal(t, "4", a.Data.CommitteeIndex)
 		}
 	})
+}
+
+func TestServer_SubmitAttestations_Ok(t *testing.T) {
+	ctx := context.Background()
+
+	transition.SkipSlotCache.Disable()
+	defer transition.SkipSlotCache.Enable()
+
+	params.SetupTestConfigCleanup(t)
+	c := params.BeaconConfig().Copy()
+	// Required for correct committee size calculation.
+	c.SlotsPerEpoch = 1
+	params.OverrideBeaconConfig(c)
+
+	_, keys, err := util.DeterministicDepositsAndKeys(1)
+	require.NoError(t, err)
+	validators := []*ethpbv1alpha1.Validator{
+		{
+			PublicKey: keys[0].PublicKey().Marshal(),
+			ExitEpoch: params.BeaconConfig().FarFutureEpoch,
+		},
+	}
+	bs, err := util.NewBeaconState(func(state *ethpbv1alpha1.BeaconState) error {
+		state.Validators = validators
+		state.Slot = 1
+		state.PreviousJustifiedCheckpoint = &ethpbv1alpha1.Checkpoint{
+			Epoch: 0,
+			Root:  bytesutil.PadTo([]byte("sourceroot1"), 32),
+		}
+		return nil
+	})
+	require.NoError(t, err)
+	b := bitfield.NewBitlist(1)
+	b.SetBitAt(0, true)
+
+	sourceCheckpoint := &ethpbv1.Checkpoint{
+		Epoch: 0,
+		Root:  bytesutil.PadTo([]byte("sourceroot1"), 32),
+	}
+	att1 := &ethpbv1.Attestation{
+		AggregationBits: b,
+		Data: &ethpbv1.AttestationData{
+			Slot:            0,
+			Index:           0,
+			BeaconBlockRoot: bytesutil.PadTo([]byte("beaconblockroot1"), 32),
+			Source:          sourceCheckpoint,
+			Target: &ethpbv1.Checkpoint{
+				Epoch: 0,
+				Root:  bytesutil.PadTo([]byte("targetroot1"), 32),
+			},
+		},
+		Signature: make([]byte, 96),
+	}
+	att2 := &ethpbv1.Attestation{
+		AggregationBits: b,
+		Data: &ethpbv1.AttestationData{
+			Slot:            0,
+			Index:           0,
+			BeaconBlockRoot: bytesutil.PadTo([]byte("beaconblockroot2"), 32),
+			Source:          sourceCheckpoint,
+			Target: &ethpbv1.Checkpoint{
+				Epoch: 0,
+				Root:  bytesutil.PadTo([]byte("targetroot2"), 32),
+			},
+		},
+		Signature: make([]byte, 96),
+	}
+
+	for _, att := range []*ethpbv1.Attestation{att1, att2} {
+		sb, err := signing.ComputeDomainAndSign(
+			bs,
+			slots.ToEpoch(att.Data.Slot),
+			att.Data,
+			params.BeaconConfig().DomainBeaconAttester,
+			keys[0],
+		)
+		require.NoError(t, err)
+		sig, err := bls.SignatureFromBytes(sb)
+		require.NoError(t, err)
+		att.Signature = sig.Marshal()
+	}
+
+	broadcaster := &p2pMock.MockBroadcaster{}
+	chainService := &blockchainmock.ChainService{State: bs}
+	s := &Server{
+		HeadFetcher:       chainService,
+		ChainInfoFetcher:  chainService,
+		AttestationsPool:  attestations.NewPool(),
+		Broadcaster:       broadcaster,
+		OperationNotifier: &blockchainmock.MockOperationNotifier{},
+	}
+
+	_, err = s.SubmitAttestations(ctx, &ethpbv1.SubmitAttestationsRequest{
+		Data: []*ethpbv1.Attestation{att1, att2},
+	})
+	require.NoError(t, err)
+	assert.Equal(t, true, broadcaster.BroadcastCalled)
+	assert.Equal(t, 2, len(broadcaster.BroadcastAttestations))
+	expectedAtt1, err := att1.HashTreeRoot()
+	require.NoError(t, err)
+	expectedAtt2, err := att2.HashTreeRoot()
+	require.NoError(t, err)
+	actualAtt1, err := broadcaster.BroadcastAttestations[0].HashTreeRoot()
+	require.NoError(t, err)
+	actualAtt2, err := broadcaster.BroadcastAttestations[1].HashTreeRoot()
+	require.NoError(t, err)
+	for _, r := range [][32]byte{actualAtt1, actualAtt2} {
+		assert.Equal(t, true, reflect.DeepEqual(expectedAtt1, r) || reflect.DeepEqual(expectedAtt2, r))
+	}
+
+	require.Equal(t, 2, s.AttestationsPool.UnaggregatedAttestationCount())
+}
+
+func TestServer_SubmitAttestations_ValidAttestationSubmitted(t *testing.T) {
+	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
+
+	transition.SkipSlotCache.Disable()
+	defer transition.SkipSlotCache.Enable()
+
+	params.SetupTestConfigCleanup(t)
+	c := params.BeaconConfig().Copy()
+	// Required for correct committee size calculation.
+	c.SlotsPerEpoch = 1
+	params.OverrideBeaconConfig(c)
+
+	_, keys, err := util.DeterministicDepositsAndKeys(1)
+	require.NoError(t, err)
+	validators := []*ethpbv1alpha1.Validator{
+		{
+			PublicKey: keys[0].PublicKey().Marshal(),
+			ExitEpoch: params.BeaconConfig().FarFutureEpoch,
+		},
+	}
+	bs, err := util.NewBeaconState(func(state *ethpbv1alpha1.BeaconState) error {
+		state.Validators = validators
+		state.Slot = 1
+		state.PreviousJustifiedCheckpoint = &ethpbv1alpha1.Checkpoint{
+			Epoch: 0,
+			Root:  bytesutil.PadTo([]byte("sourceroot1"), 32),
+		}
+		return nil
+	})
+
+	require.NoError(t, err)
+
+	sourceCheckpoint := &ethpbv1.Checkpoint{
+		Epoch: 0,
+		Root:  bytesutil.PadTo([]byte("sourceroot1"), 32),
+	}
+	b := bitfield.NewBitlist(1)
+	b.SetBitAt(0, true)
+	attValid := &ethpbv1.Attestation{
+		AggregationBits: b,
+		Data: &ethpbv1.AttestationData{
+			Slot:            0,
+			Index:           0,
+			BeaconBlockRoot: bytesutil.PadTo([]byte("beaconblockroot1"), 32),
+			Source:          sourceCheckpoint,
+			Target: &ethpbv1.Checkpoint{
+				Epoch: 0,
+				Root:  bytesutil.PadTo([]byte("targetroot1"), 32),
+			},
+		},
+		Signature: make([]byte, 96),
+	}
+	attInvalidSignature := &ethpbv1.Attestation{
+		AggregationBits: b,
+		Data: &ethpbv1.AttestationData{
+			Slot:            0,
+			Index:           0,
+			BeaconBlockRoot: bytesutil.PadTo([]byte("beaconblockroot2"), 32),
+			Source:          sourceCheckpoint,
+			Target: &ethpbv1.Checkpoint{
+				Epoch: 0,
+				Root:  bytesutil.PadTo([]byte("targetroot2"), 32),
+			},
+		},
+		Signature: make([]byte, 96),
+	}
+
+	// Don't sign attInvalidSignature.
+	sb, err := signing.ComputeDomainAndSign(
+		bs,
+		slots.ToEpoch(attValid.Data.Slot),
+		attValid.Data,
+		params.BeaconConfig().DomainBeaconAttester,
+		keys[0],
+	)
+	require.NoError(t, err)
+	sig, err := bls.SignatureFromBytes(sb)
+	require.NoError(t, err)
+	attValid.Signature = sig.Marshal()
+
+	broadcaster := &p2pMock.MockBroadcaster{}
+	chainService := &blockchainmock.ChainService{State: bs}
+	s := &Server{
+		HeadFetcher:       chainService,
+		ChainInfoFetcher:  chainService,
+		AttestationsPool:  attestations.NewPool(),
+		Broadcaster:       broadcaster,
+		OperationNotifier: &blockchainmock.MockOperationNotifier{},
+	}
+
+	_, err = s.SubmitAttestations(ctx, &ethpbv1.SubmitAttestationsRequest{
+		Data: []*ethpbv1.Attestation{attValid, attInvalidSignature},
+	})
+	require.ErrorContains(t, "One or more attestations failed validation", err)
+	expectedAtt, err := attValid.HashTreeRoot()
+	require.NoError(t, err)
+	assert.Equal(t, true, broadcaster.BroadcastCalled)
+	require.Equal(t, 1, len(broadcaster.BroadcastAttestations))
+	broadcastRoot, err := broadcaster.BroadcastAttestations[0].HashTreeRoot()
+	require.NoError(t, err)
+	require.DeepEqual(t, expectedAtt, broadcastRoot)
+
+	require.Equal(t, 1, s.AttestationsPool.UnaggregatedAttestationCount())
+}
+
+func TestServer_SubmitAttestations_InvalidAttestationGRPCHeader(t *testing.T) {
+	ctx := grpc.NewContextWithServerTransportStream(context.Background(), &runtime.ServerTransportStream{})
+
+	transition.SkipSlotCache.Disable()
+	defer transition.SkipSlotCache.Enable()
+
+	params.SetupTestConfigCleanup(t)
+	c := params.BeaconConfig().Copy()
+	// Required for correct committee size calculation.
+	c.SlotsPerEpoch = 1
+	params.OverrideBeaconConfig(c)
+
+	_, keys, err := util.DeterministicDepositsAndKeys(1)
+	require.NoError(t, err)
+	validators := []*ethpbv1alpha1.Validator{
+		{
+			PublicKey: keys[0].PublicKey().Marshal(),
+			ExitEpoch: params.BeaconConfig().FarFutureEpoch,
+		},
+	}
+	bs, err := util.NewBeaconState(func(state *ethpbv1alpha1.BeaconState) error {
+		state.Validators = validators
+		state.Slot = 1
+		state.PreviousJustifiedCheckpoint = &ethpbv1alpha1.Checkpoint{
+			Epoch: 0,
+			Root:  bytesutil.PadTo([]byte("sourceroot1"), 32),
+		}
+		return nil
+	})
+
+	require.NoError(t, err)
+
+	b := bitfield.NewBitlist(1)
+	b.SetBitAt(0, true)
+	att := &ethpbv1.Attestation{
+		AggregationBits: b,
+		Data: &ethpbv1.AttestationData{
+			Slot:            0,
+			Index:           0,
+			BeaconBlockRoot: bytesutil.PadTo([]byte("beaconblockroot2"), 32),
+			Source: &ethpbv1.Checkpoint{
+				Epoch: 0,
+				Root:  bytesutil.PadTo([]byte("sourceroot2"), 32),
+			},
+			Target: &ethpbv1.Checkpoint{
+				Epoch: 1,
+				Root:  bytesutil.PadTo([]byte("targetroot2"), 32),
+			},
+		},
+		Signature: nil,
+	}
+
+	chain := &blockchainmock.ChainService{State: bs}
+	broadcaster := &p2pMock.MockBroadcaster{}
+	s := &Server{
+		ChainInfoFetcher:  chain,
+		AttestationsPool:  attestations.NewPool(),
+		Broadcaster:       broadcaster,
+		OperationNotifier: &blockchainmock.MockOperationNotifier{},
+		HeadFetcher:       chain,
+	}
+
+	_, err = s.SubmitAttestations(ctx, &ethpbv1.SubmitAttestationsRequest{
+		Data: []*ethpbv1.Attestation{att},
+	})
+	require.ErrorContains(t, "One or more attestations failed validation", err)
+	sts, ok := grpc.ServerTransportStreamFromContext(ctx).(*runtime.ServerTransportStream)
+	require.Equal(t, true, ok, "type assertion failed")
+	md := sts.Header()
+	v, ok := md[strings.ToLower(grpcutil.CustomErrorMetadataKey)]
+	require.Equal(t, true, ok, "could not retrieve custom error metadata value")
+	assert.DeepEqual(
+		t,
+		[]string{"{\"failures\":[{\"index\":0,\"message\":\"Incorrect attestation signature: could not create signature from byte slice: signature must be 96 bytes\"}]}"},
+		v,
+	)
 }
